+++ Billbee.Api.Client/Model/OrderShipment.cs
@@ -13,7 +13,7 @@
         /// <summary>
         /// The id of the order, this shipment should be attached to
         /// </summary>
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
 
         /// <summary>
         /// A note, that should be attached to the shipment
@@ -23,11 +23,11 @@
         /// <summary>
         /// If applicable, that internal id of the provider, the shipment was send with.
         /// </summary>
-        public int? ShippingProviderId { get; set; }
+        public long? ShippingProviderId { get; set; }
 
         /// <summary>
         /// If applicable, the internal id of the shipping product, the shipment uses
         /// </summary>
-        public int? ShippingProviderProductId { get; set; }
+        public long? ShippingProviderProductId { get; set; }
     }
 }
